requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.2.0
backcall==0.2.0
decorator==5.1.1
Django==4.1.3
django-admin-interface==0.23.0
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-debug-toolbar==3.7.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==2.0.0
django-mptt==0.14.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
executing==1.2.0
ipython==8.7.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.33
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.6.0
wcwidth==0.2.5
